// n-backend/apps/api/user/src/core/addresses/interfaces/addresses.interface.ts
export interface FindAddressApiResponseObject {
  addressId: string;
  isFavorite?: boolean;
  title?: string;
  firstName?: string;
  lastName?: string;
  firstNameKana?: string;
  lastNameKana?: string;
  zipCode?: string;
  prefecture?: string;
  address1?: string;
  address2?: string;
  address3?: string;
  phone?: string;
  phone2?: string;
  email?: string;
  companyName?: string;
  departmentName?: string;
  memo?: string;
}